+++ @geogebra/solver-sdk/src/renderer/index.ts
@@ -2,4 +2,5 @@ export type { LatexSettings, LatexTransformer, MathWords } from './tree-to-latex
 export { treeToLatex, defaultMathWords } from './tree-to-latex';
 export { setsSolutionFormatter, simpleSolutionFormatter } from './solution-formatter';
 export { treeToSolver } from './tree-to-solver';
+export { treeToGgb } from './tree-to-ggb';
 export { specialTransformationLatex } from './transformation-to-latex';