+++ api/src/main/kotlin/server/api/TransformationModeller.kt
@@ -20,8 +20,8 @@ class TransformationModeller(val format: Format) {
             type = trans.type.toString(),
             tags = trans.tags?.map { it.toString() },
             path = trans.fromExpr.path.toString(),
-            fromExpr = modelExpression(trans.fromExpr),
-            toExpr = modelExpression(trans.toExpr),
+            fromExpr = format.modelExpression(trans.fromExpr),
+            toExpr = format.modelExpression(trans.toExpr),
             pathMappings = modelPathMappings(trans.toExpr.mergedPathMappings(trans.fromExpr.path!!)),
             explanation = trans.explanation?.let { modelMetadata(it) },
             formula = trans.formula?.let { modelMappedExpression(it) },
@@ -36,7 +36,7 @@ class TransformationModeller(val format: Format) {
     private fun modelTask(task: engine.steps.Task): Task {
         return Task(
             taskId = task.taskId,
-            startExpr = modelExpression(task.startExpr),
+            startExpr = format.modelExpression(task.startExpr),
             pathMappings = modelPathMappings(task.startExpr.mergedPathMappings(RootPath(task.taskId))),
             explanation = task.explanation?.let { modelMetadata(it) },
             steps = if (task.steps.isEmpty()) null else task.steps.map { modelTransformation(it) },
@@ -98,16 +98,16 @@ class TransformationModeller(val format: Format) {
 
     private fun modelMappedExpression(expr: engine.expressions.Expression): MappedExpression {
         return MappedExpression(
-            expression = modelExpression(expr),
+            expression = format.modelExpression(expr),
             pathMappings = modelPathMappings(expr.mergedPathMappings(RootPath())),
         )
     }
+}
 
-    private fun modelExpression(expr: engine.expressions.Expression): Any {
-        return when (format) {
-            Format.Latex -> expr.toLatexString()
-            Format.Solver -> expr.toString()
-            Format.Json2 -> expr.toJson()
-        }
+internal fun Format.modelExpression(expr: engine.expressions.Expression): Any {
+    return when (this) {
+        Format.Latex -> expr.toLatexString()
+        Format.Solver -> expr.toString()
+        Format.Json2 -> expr.toJson()
     }
 }