+++ engine/src/main/kotlin/engine/graphing/Point.kt
@@ -1,0 +1,6 @@
+package engine.graphing
+
+/**
+ * A 2D point in the Cartesian plane.
+ */
+data class Point(val x: Double, val y: Double)